refactor: cookie 세팅 로직 -> utils 로 빼기
